pass closure action to save author and library
